updated files

+++ postgres.sh
@@ -12,36 +12,32 @@ postgres_create_role() {
 postgres_update_pg_hba() {
     ##### Add docker and your public ip in pg_hba.conf file
     
-    # using awk
-    awk -v spip="$server_public_ip" '
-    BEGIN {
-        print "# Entries for odoosaas";
-        print "host    all     all     127.0.0.1/32    md5";
-        print "host    all     all     172.17.0.0/16    md5";
-        print "host    all     all     " spip "/32    md5";
-        print "";
+    # Check if already present
+    entry_exists() {
+        grep -qw $server_public_ip/32 $postgres_pg_hba_conf_path && \
+        grep -qw '172.17.0.0/16' $postgres_pg_hba_conf_path && \
+        grep -qw '127.0.0.1/32' $postgres_pg_hba_conf_path
     }
-    {
-        print
-    }' "$postgres_pg_hba_conf_path" > temp && mv temp "$postgres_pg_hba_conf_path"
     
-    echo "Updated pg_hba.conf file"
+    if ! entry_exists; then
+        awk -v spip="$server_public_ip" '
+            BEGIN {
+                print "# Entries for odoosaas";
+                print "host    all     all     127.0.0.1/32    md5";
+                print "host    all     all     172.17.0.0/16    md5";
+                print "host    all     all     " spip "/32    md5";
+                print "";
+            }
+            {
+                print
+        }' "$postgres_pg_hba_conf_path" > temp && mv temp "$postgres_pg_hba_conf_path"
+        
+        echo "Updated pg_hba.conf file"
+    else
+        echo "All enteries are alredy present"
+    fi
+    
     head $postgres_pg_hba_conf_path
-
-    # using sed
-    # sed -i '1i\host    all     all     127.0.0.1/32    md5' $postgres_pg_hba_conf_path
-    # echo "host    all             all             172.17.0.0/16            md5" >> $postgres_pg_hba_conf_path
-    # echo "host    all             all             "$server_public_ip"/32            md5" >> $postgres_pg_hba_conf_path
-    
-    # using nano
-    # nano $postgres_pg_hba_conf_path
-    #########################################
-    # Add these lines in ipv4 section
-    # In there line you should add your public IP Address
-    # host    all             all             127.0.0.1/32            md5
-    # host    all             all             172.17.0.0/16           md5
-    # host    all             all             $server_public_ip/32       md5
-    #########################################
     
     ##### Restart postgresql service
     systemctl restart postgresql
@@ -56,20 +52,12 @@ postgres_upadate_postgres_conf() {
     -e 's/^#*\s*max_connections\s*=.*/max_connections = 1500/' \
     $postgres_postgresql_conf_path
     
-    # using nano
-    # nano $postgres_postgresql_conf_path
-    #########################################
-    # Edit these
-    # listen_addresses = '*'
-    # max_connections = 1500
-    #########################################
-    
     ##### Restart postgresql service
     systemctl restart postgresql
-
+    
     echo "Updated postgresql.conf file"
-    cat $postgres_postgresql_conf_path | grep listen_addresses
-    cat $postgres_postgresql_conf_path | grep max_connections
+    grep listen_addresses $postgres_postgresql_conf_path
+    grep max_connections $postgres_postgresql_conf_path
     
 }
